Sea-Level Rise Team Code/requirements.txt
certifi==2020.6.20
chardet==3.0.4
gmplot==1.4.0
GrabzIt==3.4.5.1
idna==2.9
imageio==2.8.0
numpy==1.19.0
pandas==1.0.5
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
Shapely==1.7.0
six==1.15.0
urllib3==1.25.9
